embeddings/requirements.embeddings.txt
fastapi==0.135.3
langchain==1.2.14
langchain_core==1.2.25
langchain_mistralai==1.1.2
langchain_qdrant==1.1.0
langgraph==1.1.4
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.33.1
uvicorn==0.42.0
jinja2==3.1.4